==> 6_PullAntidepressantDates.R <==
# Description: This code pulls out the dates of antidepressants prescribed

# Input: the output of 5_PullAntidepressants.v4.R
library("lubridate")
filepath <- "~/MED_ORDERS.csv"
w <- read.csv(filepath)

# Pull in the patients that passed the 2/30/180 rule
filepath <- "~/4_TwoThirty180_PassFail.csv"
eMERGE_STUDYIDs <- read.csv(filepath)
eMERGE_STUDYIDs_2 <- eMERGE_STUDYIDs[(which(eMERGE_STUDYIDs[,2]==1)),1]

# Input the drug list (with or without tentatives)
checkdrugpresence <- read.csv("~/5_TotalAntidepressantPresenceAcrossPatients.csv")
incl_meds <- checkdrugpresence[,1]

# Make a matrix to output antidepressant ordered
drugs <- matrix(0, nrow=length(eMERGE_STUDYIDs_2), ncol=(length(incl_meds)+1))

# Name the column based on antidepressant
colnames(drugs) <- c("STUDY_ID", incl_meds)

# Name the rows based on STUDY_ID
drugs[,1] <- eMERGE_STUDYIDs_2
rownames(drugs) <- eMERGE_STUDYIDs_2

# Only include STUDY_ID in Orders that pass 2/30/180 rule
filteredw <- w[which(w$STUDY_ID %in% rownames(drugs)),]
filteredw[,4] <- as.character(filteredw[,4])

# Pull in file that has the patients that were prescribed antidepressants
filepath <- "~/5_AntidepressantPresence.csv"
mdd_drugs <- read.csv(filepath, header=T)

eMERGE_STUDYIDs_2 <- data.frame(mdd_drugs)

# Remove the X's and "."s from ICD-9 codes
colnames(eMERGE_STUDYIDs_2) <- c("STUDY_ID", incl_meds)

# Remove rows that have no anidepressants prescribed (261 individuals, or 0.4%)
eMERGE_STUDYIDs_2 <- eMERGE_STUDYIDs_2[-which(rowSums(eMERGE_STUDYIDs_2[2:dim(eMERGE_STUDYIDs_2)[2]])==0),]

# Create a vector of repeated pseudo dates to later be replaced by real dates 
date_fill <- rep(as.Date("1999-01-01"),(max(eMERGE_STUDYIDs_2[,2:dim(eMERGE_STUDYIDs_2)[2]])*dim(eMERGE_STUDYIDs_2)[1]))

# Create a data frame that can hold 747 dates for each patient
mdd_drugdates <- data.frame(eMERGE_STUDYIDs_2[,1],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                       date_fill[1:dim(eMERGE_STUDYIDs_2)[1]])

# Create a vector that creates "a"s to be later replaced by antidepressants that patients were prescribed.
drug_fill <- rep("a",(max(eMERGE_STUDYIDs_2[,2:dim(eMERGE_STUDYIDs_2)[2]])*dim(eMERGE_STUDYIDs_2)[1]))

# Create a data frame that can hold 747 drugs that patients were prescribed.
mdd_drugs <- data.frame(eMERGE_STUDYIDs_2[,1],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                        drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]], 
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]],
                                drug_fill[1:dim(eMERGE_STUDYIDs_2)[1]])
                            

# Colnames for just the antidepressant dates
colnames(mdd_drugdates) <- c("STUDY_ID", paste0("Date", 1:(372*2)))
colnames(mdd_drugs) <- c("STUDY_ID", paste0("Drug", 1:(372*2)))

# Drug names have to be changed into a character vector.
mdd_drugs[,2] <- as.character(mdd_drugs[,2])

# Store dates of depression codes 
for (i in seq_along(1:dim(eMERGE_STUDYIDs_2)[1])){
  check <- which(eMERGE_STUDYIDs_2[i,2:dim(eMERGE_STUDYIDs_2)[2]]>0)
  all_dates <- data.frame(date_fill[1:(372*2)])
  all_drugs <- data.frame(drug_fill[1:(372*2)])

  for (j in seq_along(1:length(check))){
    dates <- mdy(filteredw[which(filteredw$STUDY_ID==eMERGE_STUDYIDs_2[i,1]),2][grep(colnames(eMERGE_STUDYIDs_2)[check[j]+1], filteredw[which(filteredw$STUDY_ID==eMERGE_STUDYIDs_2[i,1]),4], fixed=T)])
    drugs <- filteredw[which(filteredw$STUDY_ID==eMERGE_STUDYIDs_2[i,1]),4][grep(colnames(eMERGE_STUDYIDs_2)[check[j]+1], filteredw[which(filteredw$STUDY_ID==eMERGE_STUDYIDs_2[i,1]),4], fixed=T)]
    all_dates[(which(all_dates=="1999-01-01")[1]):((which(all_dates=="1999-01-01")[1]+length(dates)-1)),1] <- ymd(dates)
    all_drugs[,1] <- as.character(all_drugs[,1])
    all_drugs[(which(all_drugs=="a")[1]):((which(all_drugs=="a")[1]+length(drugs)-1)),1] <- drugs
    if (j==length(check)){
      all_dates_todoc <- all_dates[which(all_dates[,1]!="1999-01-01"),1]
      all_drugs_todoc <- all_drugs[which(all_drugs[,1]!="a"),1]
      for (z in seq_along(1:length(all_dates_todoc))){
        mdd_drugs[,z+1] <- as.character(mdd_drugs[,z+1])
        mdd_drugs[i,z+1] <- as.character(mdd_drugs[i,z+1])
        mdd_drugs[i,z+1] <- all_drugs_todoc[z]
        mdd_drugdates[i,z+1] <- all_dates_todoc[z]
      }
    }
  }
  print(paste(i, "/", dim(eMERGE_STUDYIDs_2)[1]))
  mdd_drugs_append <- mdd_drugs[i,]
  mdd_dates_append <- mdd_drugdates[i,]
  if (i==1){
    # Create table that holds the dates of antidepressants that patients were prescribed.
    write.table(mdd_dates_append,  
                file="~/6_AntidepressantDates.csv",
                sep=",",
                append=F,
                col.names = T,
                row.names=F)
    # Create table that holds the antidepressants that patients were prescribed.
    write.table(mdd_drugs_append,  
                file="~/6_AntidepressantDrugs.csv",
                sep=",",
                append=F,
                col.names = T,
                row.names=F)
  } else {
    # Append to the table that holds the dates of antidepressants that patients were prescribed.
    write.table(mdd_dates_append,  
                file="~/6_AntidepressantDates.csv", 
                append = T, 
                sep=',', 
                row.names=F, 
                col.names=F )
    # Append to the table that holds the antidepressants that patients were prescribed.
    write.table(mdd_drugs_append,  
                file="~/6_AntidepressantDrugs.csv", 
                append = T, 
                sep=',', 
                row.names=F, 
                col.names=F )
  }
}
