==> 7_OrderedAntidepressantsandDates.R <==
## Description: This code orders the antidepressants prescribed according to the dates prescribed. 
# The code puts the antidepressants in the order for which they were prescribed, and creates a corresponding
# file that holds the dates for each prescription.

require(lubridate)
filepath1 <- "~/6_AntidepressantDrugs.csv"
mdd_drugs <- read.csv(filepath1)
filepath2 <- "~/6_AntidepressantDates.csv"
mdd_dates <- read.csv(filepath2, header=T, sep=",")

colnames(mdd_drugs) <- c("STUDY_ID", paste0("Drug", 1:744))
colnames(mdd_dates) <- c("STUDY_ID", paste0("Date", 1:744))

filepath1 <- "~/3_DepressionCodes.csv"
mdd_codes <- read.csv(filepath1)
filepath2 <- "~/3_DepressionDates.csv"
mdd_codes_dates <- read.csv(filepath2, header=T, sep=",")

colnames(mdd_codes) <- c("STUDY_ID", paste0("Code", 1:1488))
colnames(mdd_codes_dates) <- c("STUDY_ID", paste0("Date", 1:1488))

# Create vector of "a"s to later be replaced by antidepressants
drug_fill <- rep("a",(max(mdd_drugs[,2:dim(mdd_drugs)[2]])*dim(mdd_drugs)[1]))

# Create vector of psuedo dates to later be replaced by antidepressants
date_fill <- rep(as.Date("1999-01-01"),(max(mdd_drugs[,2:dim(mdd_drugs)[2]])*dim(mdd_drugs)[1]))

# Create a data frame that will hold the ordered dates
mdd_dates_ordered <- data.frame(mdd_drugs[,1],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]], 
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]],
                            date_fill[1:dim(mdd_drugs)[1]])

# Create a data frame that will hold the ordered drugs
mdd_drugs_ordered <- data.frame(mdd_drugs[,1],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]], 
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]],
                        drug_fill[1:dim(mdd_drugs)[1]])

colnames(mdd_drugs_ordered) <- c("STUDY_ID", paste0("Drug", 1:744))
colnames(mdd_dates_ordered) <- c("STUDY_ID", paste0("Date", 1:744))

for (i in seq_along(1:dim(mdd_drugs_ordered)[2])){
  mdd_drugs_ordered[,i] <- as.character(mdd_drugs_ordered[,i])
}

for (i in seq_along(1:dim(mdd_drugs)[1])){
  actualdates <- which(mdd_dates[i,]!="1999-01-01")[-1]
  leng <- length(actualdates)
  # Add dates
  mdd_dates_ordered[i,2:(leng+1)] <- mdd_dates[i,actualdates][order(mdd_dates[i,actualdates])]
  # Add drugs
    mdd_drugs_ordered[i,2:(leng+1)] <- as.character(unlist(mdd_drugs[i,actualdates][order(mdd_dates[i,actualdates])], use.names=F))
  print(paste(i, "/", dim(mdd_drugs)[1]))

}

# Write file that holds the prescribed drugs in order
write.csv(mdd_drugs_ordered, file="~/7_AntidepressantDrugs_Ordered.csv")
# Write file that holds the ordered prescription dates 
write.csv(mdd_dates_ordered, file="~/7_AntidepressantDates_Ordered.csv")
